renamed newObs to Obsidian
