+++ mod.ts
@@ -1,4 +1,4 @@
-import { ObsidianRouter } from './src/newObsidian.ts';
+import { ObsidianRouter } from './src/Obsidian.ts';
 import gql from 'https://deno.land/x/oak_graphql@0.6.2/graphql-tag/index.ts';
 
 export { ObsidianRouter, gql };
